Add timer overlay window
This is just a simple window which I'm going to render the current countdown time in. Eventually, this'll be optional and be more aesthetically pleasing - but I just wanted something quick for the time being, as I've not yet done the system tray icon showing the countdown - so I need something to render the current countdown.

#17