Add database functionality
This uses PouchDb with the IndexedDB adapter. Currently not writing any useful data - just the basic plumbing to test PouchDB out, as I haven't used it before. Had to create a dummy browser window so that the database is accessed from a render process, not a main process - as otherwise IndexedDB doesn't work.

Resolved #13