Switch database from PouchDB to NeDB (also persist Slack tokens)
This was because PounchDB required the database code to be run in the render threads, which was annoying. NeDB allows running from the main thread or the render threads. Most of the time, I want it to run from the main thread.

Resolves #52

Also, whilst testing NeDB - we're also storing the Slack authtokens. Note that at this stage, they're being stored in plain text. Need to look into the best way of storing them. Issue #16 related to Slack integration.