Add GitReleaseManager.yml
I want to support issues with the label 'Refactor' being in the milestone, which will then be ignored in the output. Currently if an issue doesn't have either 'Enhancement', 'Feature', or 'Bug' - GRM will fail.

I'm unsure if I need to explicitly reference this yaml file from the Cake script - or if it does it automatically. Will find out next time!